Add files via upload
"Create BYKToken smart contract with ERC-20 standard compliance

This commit adds the initial version of the BYKToken smart contract that implements the ERC-20 token standard. The contract includes the name, symbol, decimals, and total supply of the BYKToken. It also defines the transfer, approve, and transferFrom functions, as well as the necessary data structures to store the balances and allowances of users.

This commit lays the foundation for the development of the BYKToken, which can now be deployed and traded on the Ethereum blockchain."